digital-twin-backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
sqlalchemy==2.0.36
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.20
pandas==2.2.3
openpyxl==3.1.5
langchain==0.3.13
langchain-ollama==0.2.3
langchain-openai==0.3.1
langchain-community==0.3.13
langgraph==0.2.60
httpx==0.28.1
aiosqlite==0.20.0
